Fix/improve various small things
Fix a bug where a specific CPE would not replace another more
broad OS in the potential OS list.

Ignore a prefix like T: or U: for the validity of supplied ports.
This is done to specify which protocol the specified ports should
be scan with.

Include all potential OSes and the intermediate result into the
created files of the Nmap scan.