Change port keys of scan result
This is done to be more universal with the scan result key names.
For example name originally actually meant service, e.g. name http
meant an http server is running.

Now name is the name of the software that is running on a port and
service stores the service that is running on a port, e.g. http,
ssh, etc.