Add new config argument for CVE correlation
The new argument can make the module skip the CVE look ups for the
OS CPE(s). This argument was introduced, because OS lookups can
take fairly long, when their CPE is only general. E.g. it can
happen that the module ends up retrieving 500000 CVE entries. This
can also happen with port entries. Still, for OS entries it
happens regularly, so an option to disable it is appropriate.