Add new web crawler module
The new module uses the previous scan results and webserver_map
results to crawl / scrape a webserver. The module can discover
new URLs that are statically written into the code and URLs
that are only visited when triggering JS events, e.g. onsubmit. In
the latter case, Selenium is used to simulate user behavior. Apart
from that, the module can also discover cookies and GET / POST
parameters that are e.g. used for authentication. Additionally,
the module finds new network locations and ouputs them in a
separate result file. Finally, the module searches the crawled
pages for code comments that are also output in a module specifc,
separate result file.