Slightly improve CVE / NVD search
Before, there was the following problem that is now fixed:
When looking for NVD entries for the CPE
'cpe:/a:openbsd:openssh:7.2:p2', entries that only had
'cpe:/a:openbsd:openssh:7.2' as affected CPE were ignored b/c
the (incorrect) assumption was that e.g. here patch two (p2) would
be different from the base 7.2 version and therefore not be
affected.